Rename Util classes to *Support classes with static methods
Extract PersistenceUnitDefinitionSupport from HibernateProcessorSupport
for per-processor separation of shared logic.
