+++ extensions/hibernate-orm/deployment/src/test/java/io/quarkus/hibernate/orm/config/dialect/DialectSpecificSettingsMySQLIgnoredTest.java
@@ -11,7 +11,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
 import io.quarkus.hibernate.orm.MyEntity;
-import io.quarkus.hibernate.orm.deployment.util.HibernateProcessorUtil;
+import io.quarkus.hibernate.orm.deployment.util.HibernateProcessorSupport;
 import io.quarkus.test.QuarkusExtensionTest;
 
 public class DialectSpecificSettingsMySQLIgnoredTest {
@@ -21,11 +21,11 @@ public class DialectSpecificSettingsMySQLIgnoredTest {
             .withApplicationRoot((jar) -> jar
                     .addClass(MyEntity.class)
                     .addAsResource("application-start-offline-mysql-dialect.properties", "application.properties"))
-            .setLogRecordPredicate(record -> HibernateProcessorUtil.class.getName().equals(record.getLoggerName()))
+            .setLogRecordPredicate(record -> HibernateProcessorSupport.class.getName().equals(record.getLoggerName()))
             .overrideConfigKey("quarkus.datasource.db-kind", "") // This will override to default which is H2
             .overrideConfigKey("quarkus.hibernate-orm.dialect.mysql.bytes-per-character", "8") // This will be ignored
             .overrideConfigKey("quarkus.hibernate-orm.dialect.mysql.no-backslash-escapes", "true") // This will be ignored
-            .setLogRecordPredicate(record -> HibernateProcessorUtil.class.getName().equals(record.getLoggerName()))
+            .setLogRecordPredicate(record -> HibernateProcessorSupport.class.getName().equals(record.getLoggerName()))
             .assertLogRecords(records -> {
                 assertThat(records)
                         .extracting(LogRecord::getMessage)
